backend/requirements.txt
flask>=3.0
flask-cors>=4.0
gitpython>=3.1
networkx>=3.0
numpy>=1.24
scikit-learn>=1.3
